OBJ Loader (Nearly finished)
Custom OBJ Loader is quite messy, has a lot of comment notes for myself,
and needs quite a bit of cleanup but this all should be handled in the
next commit regarding this. The implementation is working completely
correctly.
Changed Split(...) in Globals.h and Release() in GameStateManager.cpp to
remove all memory leaks.